workbench-api-starter/requirements.txt
# KPMG Workbench API Starter - 依存関係
requests>=2.28.0
python-dotenv>=1.0.0
